+++ test/src/apps/map_actions_test.dart
@@ -19,6 +19,11 @@ void main() {
       expect(AppleMaps.directionsWithCoords(destination: coordinate), isA<MapDirectionsWithCoordsAction>());
     });
 
+    test('Citymapper actions implement supported map abstractions', () {
+      expect(Citymapper.view(coordinate: coordinate), isA<MapViewAction>());
+      expect(Citymapper.directionsWithCoords(destination: coordinate), isA<MapDirectionsWithCoordsAction>());
+    });
+
     test('Waze actions implement map abstractions', () {
       expect(Waze.view(coordinate: coordinate), isA<MapViewAction>());
       expect(Waze.search(query: 'Central Park'), isA<MapSearchAction>());